Make EntityDeleteMixin.delete deal with HTTP 204s
An HTTP 204 response indicates that a request has been processed but no content
is being returned. When this happens, any attempt to decode the JSON in a
response will fail, because no content is available to decode that content. Make
`EntityDeleteMixin.delete` deal with this case. This fixes a test failure.
Updated test results:

    $ nosetests tests/foreman/api/test_multiple_paths.py:DoubleCheckTestCase -m 1_ActivationKey
    .S.
    ----------------------------------------------------------------------
    Ran 3 tests in 31.344s

    OK (SKIP=1)