Added `enable_repos` function to ui/activationkey module.
This function helps us to enable particular repos,
which by default are disabled.

Changing default settings for content hosts that register with
this activation key requires subscription-manager version 1.10
or newer to be installed on that host.

For rhel65 the installed version is 1.9 on the clients, hence
Enabling the repo via the UI doesn't enabled the repo.

Also enabling repos uses the sub-man instead of yum-config-manager.