Fixed the UI role tests.
Updated the data for role tests accordingly, earlier data was not
right as per the test.

Also using 255 chars would not be needed, if required we can have
a separate test with a separate locator to simplify things.

The issue is the locator itself needs to be tweaked between it's
usage with locale and long char length tests.