Fix errors in class entities.ActivationKey
The paths advertised by class `ActivationKey` are incorrect. Fix this by
dropping the "subscriptions" path and adding the "add_subscriptions" and
"remove_subscriptions" paths. Update tests for method `ActivationKey.path`.

Update method `ActivationKey.add_subscriptions`. Fix a typo in the name of the
method itself, and make the method general-purpose enough to accept any set of
parameters that the URL itself accepts.