Drop nose dependency
Even though nose was listed as a optional requirement it was a real
dependency as its @attr decorator was being used in may places. The attr
was added in order to be able to run tests by feature. But using a test
runner like nose or py.test or even unittest is possible to specify
which feature to run by just running an specific test module, thanks
robottelo test organization. With that said the attr decorator is not
needed and also not wide used over the test automation.

Also move verbosity config to main section. Robottelo should have its
own verbosity configuration and should not be the same from a nose
config.