Give content view API tests better error messages
Several tests in class
`tests.foreman.api.test_contentview.CVPublishPromoteTestCase` produce the
following error message:

    AssertionError: u'error' != 'success'

This is not very useful. Make the assertions in that class produce more useful
output upon failure. This will make it easier to understand why those tests are
failing and diagnose the failures.

Several tests in this class already have more verbose error messages, but not
all of them. This commit brings quality error messages to all publish and
promote-related assertions in the class.