Fix syntax errors in CLI test code
Several test methods provide a `tuple_search` parameter to the `exists` method.
However, the `exists` method does not have a `tuple_search` argument. Its
signature looks like this:

    @classmethod
    def exists(cls, options=None, search=None):

The `search` parameter is an iterable that may contain two elements, and most
(if not all?) of the `tuple_search` arguments are two-tuples, so this is
probably what is intended. Some tests that currently fail now pass as a result
of this change. For example:

    $ nosetests tests/foreman/cli/test_org.py -m test_redmine_4486
    ......
    ----------------------------------------------------------------------
    Ran 6 tests in 67.645s

    OK