Fix failing CLI test `test_add_ptable`
Method `TestOperatingSystem.test_add_ptable` in module
`tests.foreman.cli.test_os` has been failing in Jenkins for ~100 test runs. Fix
that failure. The root cause is that a command like this is being executed:

    hammer os add-ptable --id <os_id> --ptable <ptable_name>

However, the `--ptable` flag does not exist. This accomplishes the same end goal
and is valid:

    hammer os add-ptable --id <os_id> --partition-table <ptable_name>

Make the change. Also, streamline the test itself. The test is slightly
lengthier, but should be easier to skim through due to better variable names,
better chunking of code, comments, etc.