Fix failing API tests
Fix some of the API tests that are currently failing. All fixed tests are
failing either due to extremely minor bugs that are unproblematic and unlikely
to be fixed in errata releases, or are failing due to errors in the test suite
itself. All affected tests have been tested against a Satellite 6.0.8 deploy and
pass.

Several old" API tests are failing, but have been left in place. Those failures
are meaningless. The tests themselves are broken.