Skip tests failing due to upstream BZ 1151240 fix
BZ 1151240 describes an issue wherein the API fails to provide an organization
ID in response to certain requests. This bug has apparently been fixed in
upstream, which causes the relevant tests to run again. However, the bug has not
been fixed in Satellite 6.0.x, so the now-running tests fail. Skip them. No
regression has occurred in the 6.0.x branch, and letting the tests fail gives an
incorrect impression.

    $ nosetests tests/foreman/api/test_multiple_paths.py -m test_post_and_get
    S...SS.S...S..S.
    ----------------------------------------------------------------------
    Ran 16 tests in 13.775s

    OK (SKIP=6)