Fix tests that fail due to BZ 1151240 being closed
BZ 1151240 contained two complaints:

* When one asks the server for information about an entity, the server will
  respond with incomplete information. Some foreign key relationships are not
  described.
* The format used to describe foreign key relationships is unwieldy. It's much
  easier to deal with a list of IDs than a list of hashes that (hopefully)
  contain entity IDs.

The BZ has been closed. The first complaint is allegedly addressed. The second
complaint may be addressed in APIv3, but not now. Fix tests that fail due to
this BZ being closed. It is important to note that foreign key relationships are
ignored in the fixed tests. It is very hard to validate this information given
the variable naming and unwieldy structure of foreign key relationships returned
by the server. A more productive approach for tackling this problem is to simply
flesh out all of the `read` methods.

    $ nosetests tests/foreman/api/test_multiple_paths.py -m test_put_and_get
    ......................
    ----------------------------------------------------------------------
    Ran 22 tests in 50.022s

    OK
    $ nosetests tests/foreman/api/test_multiple_paths.py -m test_post_and_get
    ......................
    ----------------------------------------------------------------------
    Ran 22 tests in 24.693s

    OK
    $ nosetests tests/foreman/api/test_architecture.py
    .S
    ----------------------------------------------------------------------
    Ran 2 tests in 0.931s

    OK (SKIP=1)