Make the `read` method work for most entities
Uncomment nearly all of the tests that call `EntityReadMixin.read`. The only
entity whose `read` method is not being tested is `ComputeResource`, due to
design issues with the class. Make all relevant tests pass:

    $ nosetests tests/foreman/api/test_multiple_paths.py:EntityReadTestCase
    .........................
    ----------------------------------------------------------------------
    Ran 25 tests in 25.024s

    OK

As a result of this change, it is now possible to execute chains of method
calls. For example, to discover the name of the organization that a repository
belongs to:

    entities.Repository(id=1).read().product.read().organization.read().name

Or, to get a list of all attributes for that organization:

    vars(entities.Repository(id=1).read().product.read().organization.read())

You can also discover some interesting aspects of the database design:

    >>> entities.System(id=some_uuid).read().organization.id == \
    ... entities.System(id=some_uuid).read().content_view.read().organization.id
    True