Fixed the syncplan tests for zstream
Why these tests where skipped?

Bcoz interval was not set and the tests fail, we had a bug for it
which was in open state for a long time now. This bug was fixed
for sat6.1 recently but not for zstream releases. Our UI automation
now uses the 'hourly' interval by default for all the test-cases.