Work around a known issue with repository deletion
A ForemanTask is launched when a repository is deleted. Unfortunately, the
task's ID is not returned, so there is no way to monitor the status of that
task. See: https://bugzilla.redhat.com/show_bug.cgi?id=1166365

This issue is unlikely to be fixed in the 6.0 branch. Insert a delay into the
repository deletion test so that the task has time to complete.

    $ nosetests tests/foreman/api/test_multiple_paths.py -m test_delete_and_get_15_Repository
    .
    ----------------------------------------------------------------------
    Ran 1 test in 26.303s

    OK