Rewrite `Entity.__init__`, make related updates
Method `Entity.__init__` is buggy. The bug lies in the fact that entity
creation behaves differently from entity reading. For example:

    >>> SomeEntity(related_entity=1).related_entity == 1
    True
    >>> SomeEntity(id=N).read().related_entity == 1
    False

This is problematic, because it is hard to know what type of object a foreign
key field is. Will a foreign key field be an integer? An `Entity` object? That's
hard to determine. Fix this issue by making `Entity.__init__` behave like
`EntityReadMixin.read`:

    >>> isinstance(SomeEntity(related_entity=1).related_entity, RelatedEntity)
    True
    >>> isinstance(SomeEntity(id=N).read().related_entity, RelatedEntity)
    True

Explain the new behaviour in `Entity`'s docstring.

Take advantage of this new behaviour and rewrite both the
`OperatingSystemEntity` and `SyncPlan` entity classes. Add a test for
`SyncPlan.read` to module `tests.foreman.api.test_multiple_paths`. Update a test
for `OperatingSystemParameter` in module
`tests.foreman.api.test_operatingsystem`.