Add basic API tests for host collections
Test whether it is possible to a create host collection that references one or
two systems. Test results:

    $ nosetests tests/foreman/api/test_hostcollection.py
    ..
    ----------------------------------------------------------------------
    Ran 2 tests in 5.679s

    OK

Interestingly, when a host collection is read back, the IDs of the systems in
that host collection are listed, but the UUIDs of the systems in that host
collection are not listed:

    {
        u'created_at': u'2015-03-06T21:40:26Z',
        u'description': None,
        u'id': 22,
        u'max_content_hosts': None,
        u'name': u'foo',
        u'organization_id': 1,
        u'permissions': {u'editable': True, u'deletable': True},
        u'system_ids': [18, 19],
        u'total_content_hosts': 2,
        u'unlimited_content_hosts': True,
        u'updated_at': u'2015-03-06T21:40:26Z',
    }

This means that it is impossible to discover which systems belong to a host
collection, as systems can only be read from the API if their UUIDs are
provided. As a result, the tests added here are not as thorough as they should
be.