Fiddle with docstrings
The module-wide docstring for `test_syncplan.py` should not state that the
module contains tests for repositories.

Drop inline code blocks of this sort:

    """Tests specific to creating new ``Sync Plans``."""

No code is being referenced, so the backticks are inappropriate. Make some
statements lower case:

    """@Test: Create a disabled Sync Plan.

"sync plan" is not a proper noun, so the capitalization is inappropriate. Join
some words:

    @Feature: Sync Plan

Most docstrings in `tests/foreman/api/` designate the feature they target by
using the name of the relevant entity class. This makes it easy to unambiguously
create consistent references to features. Thus, "SyncPlan" is a more appropriate
term.