Partially fix broken links in the documentation
When Sphinx processes a `:param:` or `:rtype:` directive, it will try to link to
the referenced object. For example, consider the following docstring snippet:

    :param dict kwargs: The keyword args supplied to :func:`request` or one of
        the convenience functions like it.
    :returns: ``True`` or ``False``
    :rtype: bool

In this case, Sphinx will try to find and link to the objects "dict" and "bool".
Unfortunately, those objects do not exist in the NailGun code base. This issue
can be masked by using the `nitpick_ignore` option in the Sphinx configuration
file, but Sphinx will still produce malformed documentation.

Partially fix this by removing unlinkable references to "list", "None" and
"string". Drop the `nitpick_ignore` option that masks errors when trying to link
to "list", thus helping to prevent this issue from creeping back in to the code
base.