Provide a default value for ComputeResource.url
Set `ComputeResource.url.default` if either "docker.internal_url" or
"docker.external_url" is set in the configuration file. This makes it possible
to work with docker compute resources as part of the test suite.

This fix only works if module `robottelo` is imported before a docker compute
resource is created. To illustrate:

    >>> from nailgun import entities
    >>> hasattr(entities.ComputeResource.url, 'default')
    False

    >>> from nailgun.entities import ComputeResource
    >>> hasattr(ComputeResource.url, 'default')
    False

    >>> import robottelo
    >>> from nailgun import entities
    >>> hasattr(entities.ComputeResource.url, 'default')
    True

    >>> import robottelo
    >>> from nailgun.entities import ComputeResource
    >>> hasattr(ComputeResource.url, 'default')
    True

This is a bit of a hack, and it will not work well in all cases. This works
because Robottelo only currently creates libvirt and docker compute resources,
and the `url` attribute has no apparent effect when creating a libvirt compute
resource. A longer-term solution is to make `ComputeResource` a base class and
have subclasses such as `LibvirtComputeResource` and `DockerComputeResource`. It
will then be possible to set `DockerComputeResource.url.default` and not affect
all types of compute resources managed by the test suite.