Update from NagilGun 0.11.0 to 0.12.0
NailGun 0.12.0 brings one significant change: missing values are no longer
created by default, and the signatures of the various `EntityCreateMixin`
methods have changed. Respond to this by overriding the default value of the new
`nailgun.entity_mixins.CREATE_MISSING` variable.

With NailGun 0.12.0 and this fix against an upstream system:

    $ nosetests tests/foreman/api/test_multiple_paths.py
    ...S...S......S...............................................................S...S..........................................................................................................................................SSSS.SSSSSS..S.S.SSSSSS...SSSSSSSSSSSSSSSSSSSSSSSSSSSSSSS
    ----------------------------------------------------------------------
    Ran 278 tests in 407.958s

    OK (SKIP=54)

With NailGun 0.12.0 but not this fix against an upstream system:

    $ nosetests tests/foreman/api/test_multiple_paths.py
    FFFSFFFSFFFFFFSFFFFFFFFFFEEEEEEEEEE^C
    …
    Ran 36 tests in 13.920s

    FAILED (SKIP=3, errors=10, failures=22)