Add test demonstrating User → Role associations
Users can be associated with roles via the API. Unfortunately, the API docs
don't make mention of this. BZ 1216239 outlines this shortcoming, and this PR
provides proof positive that users can be associated with roles via the API.
Test results against an upstream system:

    $ nosetests tests/foreman/api/test_user.py -m _with_role
    ..
    ----------------------------------------------------------------------
    Ran 2 tests in 4.558s

    OK

This test depends on the "role" field being added to the User entity. See
SatelliteQE/nailgun#57.