Update Product test_positive_delete_1 CLI test
The test is failing in automation because the info command are being
issued immediately the delete command, but after some time the product
is really deleted and the info returns the expected result.

In order to avoid this failure, add workaround to sleep for a few
seconds before running the info command but just if BZ #1219490 is open.

Closes #2147