Add helper method `get_server_software`
Method `get_server_software` logs in to the remote server, inspects it, and
guesses which version of software the server is running. Currently, it returns
one of the following strings:

* 'downstream'
* 'upstream'

Output when run against a downstream system:

    >>> from robottelo.common import helpers
    >>> helpers.get_server_software()
    'downstream'

Output when run against an upstream system:

    >>> from robottelo.common import helpers
    >>> helpers.get_server_software()
    'upstream'

Thanks @Ichimonji10 for the initial version and commit message.