Update test_bugzilla_1190122 Docker CLI test
The test is failing because the tag count is not shown in the list
output, skip the test with the corresponding bug.

Update the test to check if that the tag name have some information
BZ#1221729 states that the tag name was not displayed, because this
should be asserted for a non-zero length tag name.

Closes #2235