Fix CLI test_manifest_refresh and add a new test
Fix `test_manifest_refresh` by using a not cloned manifest instead of a
cloned one. Cloned manifests will not work since the CDN server will be
checked and the changes done are not valid.

Add a test to ensure that `hammer manifest refresh` returns a non-zero
return code if an invalid manifest (cloned one) is provided. The command
is not failing even though the API returns an error.

Closes #2352