Add click helper method for UI
The click method finds the element and then click on it. If the element
is not found then an exception is raised. This is intended to reduce the
amount of UI code needed to click on elements.

Refactor ui.base and ui.activationkey to use the new method to
illustrate how the click can save lines of code and provide better
logging if an element can not be found.

Bonus: Change double to single quote and some other code format
adjustments.