Drop unused and erroring code
The `generate_entities.py` script was used some time back to generate an initial
version of what are now NailGun's entities. The script has not been used for
some time, it does not generate entities that conform to what NailGun currently
uses, and the unit tests for the script fail with current versions of the mock
library.