Fix name length for ptable tests
2 tests which keep failing even after the SatelliteQE/nailgun#153 are `test_create_ptable_with_different_names` and `test_update_ptable_with_new_name`. They explicitly set ptable name length to randint(1, 255).
PR makes them use name of at least 4 characters long.
Fix with 4 chars length seems to be OK so far (I can't see any failure with 4 chars on Jenkins). As a bonus - updated data decorator for tests to use common lists.

I also reviewed all the partition tables calls in robottelo and nailgun, and there are no more calls with less than 4 chars - most of the calls are with create_missing, and some are with 4/6/10 chars of alpha/alphanumeric type, which should be ok.