Update docker compute resource URLs
Consider the internal docker daemon URL as the unix socket and the
external by accessing it using http. This is how the Satellite 6
recommends setting up docker as a internal compute resource (running on
the same machine as the Satellite) or a external compute resource
(running on a different machine and being accessed via http).