Drop automation-tools as a library
Automation tools was being used to provide just a little functionality
like checking for the distro information and providing the installer
test case. The former is not more necessary because the distro_info task
was being used to help validate an skip but it is not necessary anymore.
The latter is being handled by the automation when provisioning the test
machine.