Updated Docker API tests to use subTest.
Updated all Docker API tests to use `subTest` for all data-driven test
cases, and removed the `ddt` and `data` decorators.

There has not been any gain in the time it takes to run all tests but
they all passed against a Satellite 6.1.2 system, with a total of 48
tests with 1 of them skipped due to it not being implemented yet.