Updated Docker CLI tests to use subTest.
Updated all Docker CLI tests to use `subTest` for all data-driven test
cases, and removed the `ddt` and `data` decorators.

Out of the 49 tests that were executed, 17 were skipped due to the fact
that they are not yet implemented. All other tests passed and the entire
suite was run under 38 minutes.