Update CLI Import module according to #2776
The last and the only CLI module which wasn't updated according to #2776 is Import.
The reason is different from other modules approach, especially for helper functions, and I had some questions to clarify with @rplevka first.
Now everything is clear and i think it's time to make Import consistent with other modules :)