Refactor Robottelo settings
The main goal of this refactor is to make robottelo easier to be
configured. With that said, the design is focused on being easier to
check what is available on the settings object and provide useful
information for missing settings before running anything.

Now Robottelo can be configured on a selective way. Some features, like
docker or clients provisioning, require additional infrastructure to be
configured before running the tests. Robottelo identify if a feature is
available or not by checking if the related section on the settings file
is available or not.

In addition some helper methods was dropped in favor of feature related
helper methods. For example, the `settings.server` that provides some
helpers like `get_url` and `get_credentials`.

This refactor does not cover skipping tests based on the availability of
the features because this will be covered as a next step to avoid more
changes as the current changes are huge.

Finally fix documentation generation in order to leave just the warning
about the remaining classes that uses MetaCLITestCase.