Fixes `UnicodeEncodeError` which masks the real error for UI tests.
Recent automation for Operating Systems via the UI have shown the following error:

    UnicodeEncodeError: 'ascii' codec can't encode characters in position
    0-29: ordinal not in range(128)

At a first glance, this may seem to be the real issue, but the real problem is that
we're forcing utf-8 string and non-utf-8 'together'. The real culprit is this:

    TimeoutException: Timed out waiting for element
    '//a[contains(., '𢲖𣂡䱅킳쩪𣉗瀧덙䘡𣦤𫉛馰肏𦚮𦡭𑂈𩉑鹐脵𩸜𢵗ﴌ𠂩𠟟𧥪𪡼𩉚鸆𥗙㢛')]'
    to display.

Since the exception being raised tries to display the name of the entity that could
not be found, '𢲖𣂡䱅킳쩪𣉗瀧덙䘡𣦤𫉛馰肏𦚮𦡭𑂈𩉑鹐脵𩸜𢵗ﴌ𠂩𠟟𧥪𪡼𩉚鸆𥗙㢛', we end
up with a unicode error, masking the real culprit.