Add SauceLabs browser
A while back we removed support for running UI tests via SauceLabs. This
is time to get it back with extra improvements. Now one is able to
select which browser is gonna be created selenium, docker or saucelabs.

Also when choosing to run on SauceLabs, setting the webdriver will also
choose the browser that is gonna be created on SauceLabs. Valid values
are firefox, chrome and ie.