Improve UITestCase tear down
Take advantage of `addCleanup` in order to run the tear down operations:
stopping a Docker browser, quitting a Selenium browser, updating a
SauceLabs job name and status, and taking screenshot.

According to the unittest docs: "If setUp() fails, meaning that
tearDown() is not called, then any cleanup functions added will still be
called". That said robottelo is now ensuring that those operations will
run even if the `setUp()` fails, which may occur when creating dependent
entities for example.

Also add more logging in order to provide more information when a docker
container is created and stopped, when a SauceLabs job is updated and
when a screenshot is saved.