Improve UI screenshot naming
Robottelo takes screenshot just when an UI test fails and that happens just
once for every test. Because that is better to no create directories based on
the test class name and test method name. Instead put all the generated
screenshots under the same directory based on the current day and prepend the
ClassName-method_name- to the generated filename.