Drop pycrypto in favor of cryptography library (#3432)
cryptography is a new player in cryptographic world and it is really shinning
since very known libraries like Paramiko is using it now.

Before Robottelo relied on pycrypto in order to sign the cloned manifests
without listing it as a direct dependency. Drop that and add cryptography as a
new dependency.