Display Type for Compute Resource is returned in lowercase. (#3436)
Some of our Compute Resource tests for Libvirt were failing because we expected
the field `display type` to return one of the following values in uppercase:
`SPICE` or `VNC`. Turns out that the API now returns these values in lowercase.
This changeset updates the 2 failing tests to use lowercase values.