Fix UI subscription failures (#3470)
Note:
1. Before this fix, the UI subscription tests fail if the delete tests are called
first.  It is not reproducible locally.
2. I understand that this fix breaks the atomicity of tests by combining few
tests but without these changes, none of the subscription tests run in tier1 and
this reduces test coverage for subscriptions.
3. Also note that this test removes the invalid
test_positive_assert_delete_button test as it tries to assert the delete button
during the manifest delete process and this is error prone and invalid.