Introduce run_in_one_thread decorator to execute specific tests in 1 thread
`run_in_one_thread` is an optional decorator which doesn't conflict with tier decorators (both can and basically should be applied at the same time). It doesn't contain any logic on robottelo side but is used as a tag describing that specific test (or even class with tests) should be executed in 1 thread and can't be executed while some others tests are running in the background.
Together with this PR, jenkins jobs should be updated to execute 2 commands 1 by 1 - tests not decorated with run_in_one_thread in 4 threads first, and only after that tests decorated with run_in_one_thread in 1 thread. The idea is simple as that:
```bash
py.test tests/foreman -n 4 -m 'tier1 and not run_in_one_thread'
py.test tests/foreman -m 'tier1 and run_in_one_thread'
```
for each tier job accordingly.
This will allow us to easily decorate the tests which have to be executed sequentially in 1 line of code with no changes to any of jenkins jobs.