Fix random location failure with tooltip overlapping search button
When you hit 'Manage locations' link and leave cursor at that level - it will appear on the first item of the locations list. If the first item of the list has a long name and is shortened, tooltip with full
name will appear as cursor is pointing on it. And if your browser resolution is 1424px (+-100px) - tooltip will overlap search button, making it impossible to click on it.

Moving the cursor to the search button explicitly, making the tooltip disappear before actual clicking.