Change token prefix
Change the prefix from `@` to `:` this will ensure the following:

* Better compatibility with the RST format and Sphinx will produce
  better documentation.
* `make docs` will help on identifying if some of the token values are
  not formatted properly.
* Use new version of Testimony which drops the token-prefix option and
  will provide better tolling for managing the test cases' information
* Include all test cases to the docs. Update the gen_api_docs script to
  generate them

The majority of the changes are just s/@/: for the tokens but since now
following the RST format is being enforced it was needed to also update
some docstrings. The remaining changes was to update the scripts that
deal with the `:id:` token and test-docstrings make target.