we need UI tests to create compute resources in short term
Currently we do not have UI tests which would create CR and because e.g. hammer skips extra verification like "Load datacenters", it is missing some recently discovered issues. I'm enabling these tests so we have an early indication of such an issue next time. CC @alda519 

Why tier2? When CR for these providers is being created, we also click at "Load datacenters/regions". We do not click on "Load tenants", but that is IMO a place for an improvement.