Refactor assertions to make test less brittle
Counting number of lines in stdout was not a good idea to begin with.
Just check if expected docker repository URI is not in search output
(when unauthenticated pull == false), or is in search output (otherwise
and after login). Docker repository URI are required to be unique across
entire installation, so there is no risk of catching different, but
named the same repository.