Fix containers removal upgrade tests
Or, at least, make it pass. The problem is that due to changes in
registry.access.redhat.com, Satellite is no longer able to pull in
content from that registry. There is hotfix available for 6.5.z, but
it's not yet part of proper release. This "solution" simply checks if
relevant bugzilla is opened and if it is, it doesn't try to pull
content.