Fix test_positive_multi_registry
Repository synchronization operation regularly failed in this test,
because Nailgun gives only 5 minutes to complete it, while it takes
around 10 minutes to synchronize repositories in question.

Fixed by syncing only single tag instead of entire repo. The goal of test
is to check if we support various remote registries, and synchronizing
single tag is enough to tell that.