Update test_host_info, clear lru_cache in fixtures
This is causing interaction issues with other unit tests, as the lru_cache is separate from the test module

The mocked values from the test were getting cached and retreived in other test modules

This means the test_vm module will fail if run without a host configured, as it will try to ssh to localhost

FIXES #7635

Rework test_vm for pytest, failing due to host ssh

Intial rework to convert test_vm to pytest
The failures are because of the commit to resolve issue #7635, which cleared the cached host_os value that was mocked by another test module

Requires additional work to resolve the test failures. The host values should be mocked, or a container with sshd running spawned for the test to connect to as a host.

Add host os mocking to test_vm