Refactor issue_handlers to apply a single BZ marker
Register the BZ marker, and update issue_handlers to pass it a list of all the BZ's marked through skip_if_open decorator.

Add a --BZ option to maintain the current support for test case filtering on specific BZs. The argument takes a comma separated list of BZs.

Verbosely log when a testcase is deselected based on these filters and BZ marks so its obvious why a test is or isn't included in collection, and what marks apply.

Tested with existing test cases that have singular and multiple skip_if_open mark decorators applied.
Tested with singular and multiple BZ filters passed on CLI.
Collection and logging were as expected.

Further refactoring could simplify the use of the `skip_if_open` decorator, building the skip behavior into a direct BZ marker.  The tests would have to be updated with a new mark decorator, and the hook updated to not apply BZ marks, but read them to apply skipif marks, which it already is doing.